Move cat testing into subdir
