+++ jctools-core/src/test/java/org/cliffc/high_scale_lib/cat_test/LockCounter.java
@@ -1,4 +1,4 @@
-package org.cliffc.high_scale_lib;
+package org.cliffc.high_scale_lib.cat_test;
 import java.util.concurrent.locks.*;
 public final class LockCounter extends Counter {
   public String name() { return "Lock"; }
